refactor(exec): extract rate limits + sessions root resolver
